Refactored and fixes.
